♻️ refactor : fetchWineDetail 데이터 페치 함수 통일, 에러, 404, 로딩 처리 개선
