+++ src/app/(with-header)/wines/[id]/_components/ReviewContainer.tsx
@@ -1,43 +1,14 @@
 'use client';
-import { useState, useEffect } from 'react';
-import { useParams } from 'next/navigation';
-import { fetchWineDetail } from '@/lib/fetchWineDetail';
-import { ReviewData } from '@/types/review-data';
+import { useState } from 'react';
 import { calculateTasteAverage, getTopThreeAromas, calculateRatingCount } from '@/utils/ReviewUtils';
+import { ReviewData } from '@/types/review-data';
+import { EditReviewData } from '@/types/review-data';
+import { AddReviewData } from '@/types/review-data';
 import ReviewTasteAverage from './ReviewTasteAverage';
 import ReviewAroma from './ReviewAroma';
 import ReviewItem from './ReviewItem';
 import ReviewRating from './ReviewRating';
 import NoReview from './NoReview';
-
-export interface AddReviewData {
-  reviewId: number;
-  rating: number;
-  lightBold: number;
-  smoothTannic: number;
-  drySweet: number;
-  softAcidic: number;
-  aroma: string[];
-  content: string;
-  user: {
-    id: number;
-    nickname: string;
-    image: string;
-  };
-  wineId: number;
-  wineName: string;
-}
-
-export interface EditReviewData {
-  rating: number;
-  lightBold: number;
-  smoothTannic: number;
-  drySweet: number;
-  softAcidic: number;
-  aroma: string[];
-  content: string;
-  wineId: number;
-}
 
 function ReviewList({
   reviews,
@@ -60,69 +31,16 @@ function ReviewList({
   );
 }
 
-export default function ReviewContainer() {
-  const { id } = useParams();
-  const wineId = typeof id === 'string' ? Number(id) : NaN;
-  const [reviews, setReviews] = useState<ReviewData['reviews']>([]);
-  const [wineName, setWineName] = useState<string>('');
-  const [avgRating, setAvgRating] = useState<number>(0);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [myReviewData, setMyReviewData] = useState<ReviewData['reviews']>([]);
-
-  useEffect(() => {
-    const fetchReviews = async () => {
-      if (isNaN(wineId)) {
-        setError('유효한 와인 ID가 아닙니다.');
-        setLoading(false);
-        return;
-      }
-
-      setLoading(true);
-      try {
-        const data = await fetchWineDetail(wineId);
-        setWineName(data.name);
-        setReviews(data.reviews);
-        setAvgRating(data.avgRating);
-        setMyReviewData(data.reviews);
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err.message);
-        }
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    if (wineId) {
-      fetchReviews();
-    }
-  }, [wineId]);
+export default function ReviewContainer({ data }: { data: ReviewData }) {
+  const { reviews, avgRating } = data;
+  const [localReviews, setLocalReviews] = useState(reviews);
 
   const deleteMyReview = (id: number) => {
-    const updatedReviewList = myReviewData.filter((value) => value.id !== id);
-    setMyReviewData(updatedReviewList);
-
-    const updatedReviews = reviews.filter((value) => value.id !== id);
-    setReviews(updatedReviews);
+    setLocalReviews((prevReviews) => prevReviews.filter((review) => review.id !== id));
   };
 
   const editMyReview = (id: number, editReviewData: EditReviewData, updatedAt: string) => {
-    const updatedReviewList = myReviewData.map((value) => {
-      if (value.id === id) {
-        return { ...value, ...editReviewData, updatedAt: updatedAt };
-      }
-      return value;
-    });
-    setMyReviewData(updatedReviewList);
-
-    const updatedReviews = reviews.map((value) => {
-      if (value.id === id) {
-        return { ...value, ...editReviewData, updatedAt: updatedAt };
-      }
-      return value;
-    });
-    setReviews(updatedReviews);
+    setLocalReviews((prevReviews) => prevReviews.map((review) => (review.id === id ? { ...review, ...editReviewData, updatedAt } : review)));
   };
 
   const addReview = (newReview: AddReviewData) => {
@@ -131,15 +49,12 @@ export default function ReviewContainer() {
       id: newReview.reviewId,
       createdAt: new Date().toISOString(),
       updatedAt: new Date().toISOString(),
-      wine: { id: wineId, name: wineName, image: '', avgRating: 0 },
+      wine: { id: data.id, name: data.name, image: '', avgRating: 0 },
       isLiked: false,
     };
 
-    setReviews((prevReviews) => [formattedReview, ...prevReviews]);
+    setLocalReviews((prevReviews) => [formattedReview, ...prevReviews]);
   };
-
-  if (loading) return <div>Loading...</div>;
-  if (error) return <div>{error}</div>;
 
   const averages = calculateTasteAverage(reviews);
   const topThreeAromas = getTopThreeAromas(reviews);
@@ -163,7 +78,7 @@ export default function ReviewContainer() {
 
             <div className='mt-[60px] flex justify-between gap-[60px] tablet:flex-col-reverse tablet:px-6'>
               <div>
-                <ReviewList reviews={reviews} wineName={wineName} deleteMyReview={deleteMyReview} editMyReview={editMyReview} />
+                <ReviewList reviews={localReviews} wineName={data.name} deleteMyReview={deleteMyReview} editMyReview={editMyReview} />
               </div>
               <div className='relative'>
                 <div className='sticky top-28'>
