+++ src/app/(with-header)/wines/[id]/_components/ReviewRating.tsx
@@ -1,7 +1,7 @@
 'use client';
 import StaticRating from '@/components/StaticRating';
 import PostReviewModal from '@/components/modal/PostReviewModal';
-import { AddReviewData } from './ReviewContainer';
+import { AddReviewData } from '@/types/review-data';
 
 type ReviewRatingProps = {
   count: number;
